requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
weasyprint==62.3
Jinja2==3.1.4
google-cloud-storage==2.18.2